// frontend/src/app/api/models/coleccion-actualizacion.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ColeccionCampos } from '../models/coleccion-campos';
import { ContenidoComunCampos } from '../models/contenido-comun-campos';
import { VersionCampo } from '../models/version-campo';
type ColeccionActualizacion$ = ContenidoComunCampos & ColeccionCampos & VersionCampo;
type RequiredProperties = 'titulo' | 'version';
export type ColeccionActualizacion = ColeccionActualizacion$ & Required<Pick<ColeccionActualizacion$, RequiredProperties>>;