// frontend/src/app/api/models/destino-actualizacion.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CamposOperacionPublicada } from '../models/campos-operacion-publicada';
import { ContenidoComunCampos } from '../models/contenido-comun-campos';
import { DestinoCampos } from '../models/destino-campos';
import { VersionCampo } from '../models/version-campo';
type DestinoActualizacion$ = ContenidoComunCampos & DestinoCampos & VersionCampo & CamposOperacionPublicada;
type RequiredProperties = 'titulo' | 'version';
export type DestinoActualizacion = DestinoActualizacion$ & Required<Pick<DestinoActualizacion$, RequiredProperties>>;