// frontend/src/app/api/models/guia-actualizacion.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ContenidoComunCampos } from '../models/contenido-comun-campos';
import { GuiaCampos } from '../models/guia-campos';
import { VersionCampo } from '../models/version-campo';
type GuiaActualizacion$ = ContenidoComunCampos & GuiaCampos & VersionCampo;
type RequiredProperties = 'titulo' | 'version';
export type GuiaActualizacion = GuiaActualizacion$ & Required<Pick<GuiaActualizacion$, RequiredProperties>>;