// frontend/src/app/api/models/itinerario-actualizacion.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ContenidoComunCampos } from '../models/contenido-comun-campos';
import { ItinerarioCampos } from '../models/itinerario-campos';
import { VersionCampo } from '../models/version-campo';
type ItinerarioActualizacion$ = ContenidoComunCampos & ItinerarioCampos & VersionCampo;
type RequiredProperties = 'titulo' | 'version';
export type ItinerarioActualizacion = ItinerarioActualizacion$ & Required<Pick<ItinerarioActualizacion$, RequiredProperties>>;